+++ Backend/Repositories/Interfaces/IRepository.cs
@@ -5,9 +5,9 @@ namespace Backend.Repositories.Interfaces;
 public interface IRepository<T> where T : class
 {
     Task<T> Add(T obj);
-    Task<T?> GetById(long id);
+    Task<T?> GetById(int id);
     Task<IEnumerable<T>> Where(Expression<Func<T, bool>> predicate);
     Task<IEnumerable<T>> GetAll();
     Task<T> Update(T obj);
-    Task<bool> Remove(long id);
+    Task<bool> Remove(int id);
 }
